Add payment field to first step and move message to last step
