+++ src/components/one-time-donation/helpers/validation-schema.ts
@@ -2,11 +2,17 @@ import * as yup from 'yup'
 import { name, phone, email } from 'common/form/validation'
 import { FirstStep, SecondStep, ThirdStep } from 'gql/donations'
 
-export const validateFirst: yup.SchemaOf<FirstStep> = yup.object().defined().shape({
-  message: yup.string().notRequired(),
-  anonymous: yup.bool().required(),
-  amount: yup.number().required(),
-})
+export const validateFirst: yup.SchemaOf<FirstStep> = yup
+  .object()
+  .defined()
+  .shape({
+    payment: yup.string().required().oneOf(['card', 'bank']),
+    amount: yup.number().when('payment', {
+      is: 'card',
+      // Here we should fetch the possible payments to put into the oneOf, but it's not that important
+      then: yup.number().required().oneOf([500, 1000]),
+    }),
+  })
 
 export const validateSecond: yup.SchemaOf<SecondStep> = yup
   .object()
@@ -22,9 +28,7 @@ export const validateSecond: yup.SchemaOf<SecondStep> = yup
     personsPhone: phone.notRequired(),
   })
 
-export const validateThird: yup.SchemaOf<ThirdStep> = yup
-  .object()
-  .defined()
-  .shape({
-    payment: yup.string().required().oneOf(['bank'], 'errors-fields.bank-payment'),
-  })
+export const validateThird: yup.SchemaOf<ThirdStep> = yup.object().defined().shape({
+  message: yup.string().notRequired(),
+  anonymous: yup.bool().required(),
+})
